+++ packages/react/src/components/CardEditor/CardEditForm/CardEditFormItems/DataSeriesFormItemModal.jsx
@@ -37,6 +37,9 @@ import ThresholdsFormItem from './ThresholdsFormItem';
 
 const { iotPrefix } = settings;
 
+/* istanbul ignore next */
+const noop = () => {};
+
 const propTypes = {
   /* card value */
   cardConfig: PropTypes.shape({
@@ -193,12 +196,12 @@ const defaultProps = {
   isLarge: false,
   validDataItems: [],
   testId: 'aggregation-methods',
-
   actions: {
-    onEditDataItem: null,
+    onEditDataItem: noop,
     dataSeriesFormActions: {
-      hideAggregationsDropDown: null,
-      onAddAggregations: null,
+      hasAggregationsDropDown: noop,
+      hasDataFilterDropdown: noop,
+      onAddAggregations: noop,
     },
   },
 };
@@ -232,9 +235,8 @@ const DataSeriesFormItemModal = ({
   i18n,
   isLarge,
   testId,
-
   actions: {
-    dataSeriesFormActions: { hideAggregationsDropDown, onAddAggregations },
+    dataSeriesFormActions: { hasAggregationsDropDown, onAddAggregations, hasDataFilterDropdown },
   },
 }) => {
   const mergedI18n = { ...defaultProps.i18n, ...i18n };
@@ -301,7 +303,7 @@ const DataSeriesFormItemModal = ({
   const DataEditorContent = useMemo(
     () => (
       <>
-        {hideAggregationsDropDown(editDataItem) && (
+        {hasAggregationsDropDown && hasAggregationsDropDown(editDataItem) && (
           <div className={`${baseClassName}--input-group`}>
             {!initialAggregation || !isSummaryDashboard ? ( // selector should only be use-able in an instance dash or if there is no initial aggregation
               <div className={`${baseClassName}--input-group--item-half`}>
@@ -528,8 +530,8 @@ const DataSeriesFormItemModal = ({
           </div>
         )}
 
-        {isSummaryDashboard &&
-          type !== CARD_TYPES.TABLE && ( // only show data filter in summary dashboards
+        {hasDataFilterDropdown &&
+          hasDataFilterDropdown(cardConfig) && ( // only show data filter in summary dashboards or instance dashboard for DEVICE_TYPE
             <div className={`${baseClassName}--input-group ${baseClassName}--input-group--bottom `}>
               <div
                 className={classnames({
@@ -622,7 +624,8 @@ const DataSeriesFormItemModal = ({
       cardConfig,
       editDataItem,
       handleTranslation,
-      hideAggregationsDropDown,
+      hasAggregationsDropDown,
+      hasDataFilterDropdown,
       id,
       initialAggregation,
       initialGrain,