+++ packages/react/src/components/DateTimePicker/DateTimePickerV2.test.jsx
@@ -27,6 +27,7 @@ const defaultPresets = [
 const { iotPrefix, prefix } = settings;
 
 const dateTimePickerProps = {
+  testId: 'date-time-picker',
   onCancel: jest.fn(),
   onApply: jest.fn(),
 };
@@ -35,6 +36,8 @@ const i18n = {
   presetLabels: ['Last 30 minutes', 'Missed in translation'],
   intervalLabels: ['minutes', 'Missed in translation'],
   relativeLabels: ['Missed in translation'],
+  invalidText: 'The date time entered is invalid',
+  invalidLabel: 'Invalid',
 };
 
 describe('DateTimePickerV2', () => {
@@ -398,7 +401,7 @@ describe('DateTimePickerV2', () => {
     userEvent.click(screen.getAllByLabelText('Calendar')[0]);
 
     // open start time picker
-    userEvent.click(screen.getByTestId('date-time-picker-time-btn-1'));
+    fireEvent.focus(screen.getByTestId('date-time-picker-input-1'));
 
     // select hour
     userEvent.click(
@@ -416,7 +419,7 @@ describe('DateTimePickerV2', () => {
     screen.getByTestId('date-time-picker-spinner').blur();
 
     // open end time picker
-    userEvent.click(screen.getByTestId('date-time-picker-time-btn-2'));
+    fireEvent.focus(screen.getByTestId('date-time-picker-input-2'));
     // select hour
     userEvent.click(
       within(screen.getByTestId('date-time-picker-spinner-list-spinner-1')).getByText('05')
@@ -463,7 +466,7 @@ describe('DateTimePickerV2', () => {
 
     const startTime = screen.getByTestId('date-time-picker-input');
     // open time picker
-    userEvent.click(screen.getByTestId('date-time-picker-time-btn'));
+    fireEvent.focus(screen.getByTestId('date-time-picker-input'));
     // select hour
     userEvent.click(
       within(screen.getByTestId('date-time-picker-spinner-list-spinner-1')).getByText('05')
@@ -830,7 +833,7 @@ describe('DateTimePickerV2', () => {
     userEvent.click(screen.getAllByLabelText('Calendar')[0]);
 
     // open start time picker
-    userEvent.click(screen.getByTestId('date-time-picker-time-btn-1'));
+    fireEvent.focus(screen.getByTestId('date-time-picker-input-1'));
 
     // select hour
     userEvent.click(
@@ -1560,6 +1563,7 @@ describe('DateTimePickerV2', () => {
     expect(screen.getByText(i18n.applyBtnLabel)).toBeDisabled();
 
     // 2020-04-01 13:34 to 2020-04-01 11:49
+    fireEvent.focus(startTime);
     userEvent.type(
       startTime,
       '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}01:34 PM'
@@ -1570,6 +1574,7 @@ describe('DateTimePickerV2', () => {
     expect(screen.getByText(i18n.applyBtnLabel)).toBeDisabled();
 
     // 2020-04-01 13:34 to 2020-04-01 12:49
+    fireEvent.focus(endTime);
     userEvent.type(
       endTime,
       '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}12:49 PM'
@@ -1591,6 +1596,7 @@ describe('DateTimePickerV2', () => {
     expect(screen.getByText(i18n.applyBtnLabel)).toBeEnabled();
 
     // 2020-04-01 13:50 to 2020-04-01 13:49
+    fireEvent.focus(startTime);
     userEvent.type(
       startTime,
       '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}01:50 PM'
@@ -1607,6 +1613,8 @@ describe('DateTimePickerV2', () => {
     // expect(screen.getByText(i18n.applyBtnLabel)).toBeDisabled();
 
     userEvent.type(startTime, '{backspace}{backspace}{backspace}{backspace}{backspace}51 PM');
+
+    fireEvent.focus(endTime);
     userEvent.type(
       endTime,
       '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}9999'
@@ -1743,6 +1751,7 @@ describe('DateTimePickerV2', () => {
     expect(screen.getByText(i18n.applyBtnLabel)).toBeEnabled();
 
     // 2020-04-01 13:34 to 2020-04-06 11:49
+    fireEvent.focus(startTime);
     userEvent.type(
       startTime,
       '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}01:34 PM'
@@ -1753,6 +1762,7 @@ describe('DateTimePickerV2', () => {
     expect(timeRange).toHaveTextContent('2020-04-01 13:34 to 2020-04-06 11:49');
 
     // 2020-04-01 13:34 to 2020-04-06 12:49
+    fireEvent.focus(endTime);
     userEvent.type(
       endTime,
       '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}12:49 PM'
@@ -1764,6 +1774,7 @@ describe('DateTimePickerV2', () => {
     expect(screen.getByText(i18n.applyBtnLabel)).toBeEnabled();
 
     // 2020-04-01 13:34 to 2020-04-06 13:49
+    fireEvent.focus(endTime);
     userEvent.type(
       endTime,
       '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}01:49 PM'
@@ -1774,6 +1785,7 @@ describe('DateTimePickerV2', () => {
     expect(screen.getByText(i18n.applyBtnLabel)).toBeEnabled();
 
     // 2020-04-01 13:50 to 2020-04-06 13:49
+    fireEvent.focus(startTime);
     userEvent.type(
       startTime,
       '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}01:50 PM'
@@ -1783,7 +1795,10 @@ describe('DateTimePickerV2', () => {
     expect(timeRange).toHaveTextContent('2020-04-01 13:50 to 2020-04-06 13:49');
     expect(screen.getByText(i18n.applyBtnLabel)).toBeEnabled();
 
+    fireEvent.focus(startTime);
     userEvent.type(startTime, '{backspace}{backspace}{backspace}{backspace}{backspace}51 PM');
+
+    fireEvent.focus(endTime);
     userEvent.type(
       endTime,
       '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}9999'
@@ -1805,5 +1820,43 @@ describe('DateTimePickerV2', () => {
     expect(screen.getByTestId('date-time-picker').getAttribute('id')).toMatch(
       /-iot--date-time-pickerv2__wrapper/
     );
+  });
+
+  it('should show invalid text when in invalid state', () => {
+    render(<DateTimePicker {...dateTimePickerProps} i18n={i18n} invalid />);
+    expect(screen.getByText(i18n.invalidText)).toBeInTheDocument();
+    expect(
+      screen
+        .getByTestId(`${dateTimePickerProps.testId}-datepicker-flyout-button`)
+        .classList.contains(`${iotPrefix}--date-time-picker--trigger-button-invalid`)
+    ).toBe(true);
+  });
+
+  it('should show invalid text when in invalid state (new time spinner)', () => {
+    render(<DateTimePicker {...dateTimePickerProps} i18n={i18n} invalid useNewTimeSpinner />);
+    expect(screen.getByText(i18n.invalidText)).toBeInTheDocument();
+    expect(
+      screen
+        .getByTestId(`${dateTimePickerProps.testId}-datepicker-flyout-button`)
+        .classList.contains(`${iotPrefix}--date-time-picker--trigger-button-invalid`)
+    ).toBe(true);
+  });
+
+  it('should disable menu button when disabled is set', () => {
+    render(<DateTimePicker {...dateTimePickerProps} disabled />);
+    expect(
+      screen
+        .getByTestId(`${dateTimePickerProps.testId}-datepicker-flyout-button`)
+        .classList.contains(`${iotPrefix}--date-time-picker--trigger-button-disabled`)
+    ).toBe(true);
+  });
+
+  it('should disable menu button when disabled is set (new time spinner)', () => {
+    render(<DateTimePicker {...dateTimePickerProps} disabled useNewTimeSpinner />);
+    expect(
+      screen
+        .getByTestId(`${dateTimePickerProps.testId}-datepicker-flyout-button`)
+        .classList.contains(`${iotPrefix}--date-time-picker--trigger-button-disabled`)
+    ).toBe(true);
   });
 });